Split the functions into separate files
