+++ lib/catch.js
@@ -1,0 +1,5 @@
+var noop = require('./noop');
+
+module.exports = function (reject) {
+  return this.then(noop, reject);
+};
