+++ lib/defaultData.js
@@ -1,0 +1,6 @@
+module.exports = {
+  __state     : 'pending',
+  __stack     : [],
+  __data      : undefined,
+  __error     : undefined
+};
